+++ validated/src/main/java/uk/ac/standrews/cs/digitising_scotland/verisim/populationStatistics/dataDistributionTables/statsKeys/PartneringStatsKey.java
@@ -24,7 +24,7 @@ import uk.ac.standrews.cs.digitising_scotland.verisim.dateModel.timeSteps.Compou
  */
 public class PartneringStatsKey extends StatsKey {
 
-    public PartneringStatsKey(Integer age, int forNumberOfFemales, CompoundTimeUnit consideredTimePeriod, Date currentDate) {
+    public PartneringStatsKey(Integer age, double forNumberOfFemales, CompoundTimeUnit consideredTimePeriod, Date currentDate) {
         super(age, forNumberOfFemales, consideredTimePeriod, currentDate);
     }
